refactored dataset into a separate class
