+++ data/x_and_y_to_hdf5.py
@@ -1,17 +1,12 @@
 # -*- coding: utf-8 -*-
-"""X and Y to hdf5
-
-Automatically generated by Colaboratory.
-
-Original file is located at
-    https://colab.research.google.com/drive/1w3mMw7iKEXn4lDQhag6oObLWMuhOB6HD
+"""
+    Sample script for constructing a sequence dataset with DeepDish
 """
 
 import numpy as np
 import random
 import deepdish as dd
 import string
-import h5py
 from torch.utils.data import Dataset, DataLoader
 from torchvision import transforms, utils
 
@@ -20,7 +15,7 @@ X = []
 for i in range(0,N):
   rand = random.randint(0, 10)
   np_a = np.array([[rand for x in range(rand)] for y in range(rand)])
-  X.append(np_a)  
+  X.append(np_a)
 
 Y = [''.join(random.choices(string.ascii_uppercase + string.digits, k=N*5)) for i in range(0, N)]
 dd.io.save('test.h5', (X, Y))
@@ -32,20 +27,6 @@ recreation
 
 len(recreation['data'][1])
 
-class XY(Dataset):
-    """X to Y dataset."""
-
-    def __init__(self, hdf5_location):
-
-        self.hdf5 = dd.io.load(hdf5_location, '/')['data']
-
-    def __len__(self):
-        return len(self.hdf5[1])
-
-    def __getitem__(self, idx):
-        return (self.hdf5[0][idx], self.hdf5[1][idx])
-
 dataset = XY('test.h5')
 
 len(dataset), dataset[3]
-
